minor-final/requirements.txt
streamlit
openai
python-dotenv
gTTS
